% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_pvalues.R
\name{adjust_pvalues}
\alias{adjust_pvalues}
\title{Adjust p-values and perform threshold filtering based on the adjusted
p-values.}
\usage{
adjust_pvalues(result, pAdjustMethod = "bonferroni", pAdjustThreshold = 0.05)
}
\arguments{
\item{result}{Dataframe that contains gene-SNP pairs' information.}

\item{pAdjustMethod}{Methods for p-value adjusting, one of "bonferroni",
"holm", "hochberg", "hommel" or "BH". The default option is "bonferroni".}

\item{pAdjustThreshold}{Only SNP-Gene pairs with adjusted p-values meeting
the threshold will be displayed. Default by 0.05.}
}
\value{
A dataframe that has been adjusted and filtered, containing
information on gene-SNP pairs.
}
\description{
Adjust p-values and perform threshold filtering based on the adjusted
p-values.
}
\examples{
example_data <- data.frame(
  gene = c("Gene1", "Gene2", "Gene3", "Gene4"),
  SNP = c("SNP1", "SNP2", "SNP3", "SNP4"),
  pvalue = c(0.001, 0.04, 0.03, 0.0005))
pAdjustMethod <- "BH"
pAdjustThreshold <- 0.05
adjusted_result <- adjust_pvalues(example_data, pAdjustMethod,
pAdjustThreshold)
}
