% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildZINB-callQTL.R
\name{buildZINB}
\alias{buildZINB}
\title{Build a ZINB model.}
\usage{
buildZINB(counts)
}
\arguments{
\item{counts}{A numeric vector of gene expression values.}
}
\value{
A list containing four parameters estimated from the ZINB model.
}
\description{
Build a ZINB model.
}
\examples{
data(GeneData)
gene <- unlist(GeneData[1, ])
result <-buildZINB(gene)
}
