% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM_normalize-normalizeGene.R
\name{CPM_normalize}
\alias{CPM_normalize}
\title{Normalize gene expression using CPM.}
\usage{
CPM_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{A numeric matrix of raw gene expression counts,
with genes as rows and cells as columns.}
}
\value{
A normalized gene expression matrix after applying CPM normalization.
}
\description{
\code{CPM_normalize()} applies Counts Per Million (CPM) normalization to a
raw gene expression matrix.
}
\examples{
data(testGene)
CPM_normalize(testGene)
}
