% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{testGene}
\alias{testSNP}
\alias{testSeurat}
\alias{testSNP2}
\alias{testEQTL}
\title{Test Gene Expression Dataset}
\format{
A numeric matrix with 100 rows (genes) and 2705 columns (cells).

A numeric matrix with variants as rows and cells as columns.

An object of class \code{eQTLObject}.

A numeric matrix with variants as rows and cells as columns.
}
\description{
A dataset containing example gene expression data for testing purposes.
It contains 100 rows and 2705 columns. The row names represent gene IDs or
SYMBOLs, and the column names represent cell IDs.

A dataset containing single nucleotide variant data. Each row is one variant
and each column is one cell.

A Seurat object representing single-cell RNA-seq data.

A smaller dataset containing single nucleotide variant data for testing.
Each row is one variant and each column is one cell.

An \code{eQTLObject} created using \code{createQTLObject()}, where the raw
expression matrix is normalized using \code{normalizeGene()}, and both the
genotype matrix and normalized expression matrix are filtered using
\code{filterGeneSNP()}.
}
\keyword{An}
\keyword{\code{eQTLObject}.}
\keyword{class}
\keyword{datasets}
\keyword{object}
\keyword{of}
