% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_normalize-normalizeGene.R
\name{DESeq_normalize}
\alias{DESeq_normalize}
\title{Normalize the gene expression matrix with DESeq2.}
\usage{
DESeq_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{A numeric matrix of raw gene expression counts,
with genes as rows and cells as columns.}
}
\value{
A normalized gene expression matrix after applying DESeq
normalization.
}
\description{
\code{DESeq_normalize()} normalizes a raw gene expression matrix using the
\pkg{DESeq2} package.
}
\examples{
data(testGene)
DESeq_normalize(testGene)
}
