% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQTLObject.R
\docType{class}
\name{eQTLObject-class}
\alias{eQTLObject-class}
\title{Class \code{eQTLObject}}
\value{
An S4 object of class \code{eQTLObject}.
}
\description{
The eQTLObject class is designed to store data related to eQTL analysis,
including data lists, result data frames, and additional metadata such as
classification, species, and grouping information.
}
\section{Slots}{

\describe{
\item{\code{rawData}}{A gene expression dataframe, the row names represent gene IDs
and the column names represent cell IDs.}

\item{\code{filterData}}{Gene expression matrix after normalizing.}

\item{\code{eQTLResult}}{The result dataframe obtained the sc-eQTL results.}

\item{\code{biClassify}}{Logical; whether to convert genotype encoding in snpMatrix
to 0, 1, and 2. \code{TRUE} indicates conversion; \code{FALSE} indicates no
conversion (default).}

\item{\code{species}}{The species that the user wants to select, human or mouse.}

\item{\code{groupBy}}{Options for cell grouping, users can choose celltype,
cellstatus, etc., depending on metadata.}

\item{\code{useModel}}{model for fitting dataframe.}
}}

