% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_filter_data,eQTLObject-method}
\alias{get_filter_data,eQTLObject-method}
\title{Method to access eQTLObject filter data}
\usage{
\S4method{get_filter_data}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
filtered matrices.
}
\description{
Method to access eQTLObject filter data
}
