% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_filter_data}
\alias{get_filter_data}
\title{Generic to access eQTLObject filter data}
\usage{
get_filter_data(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
filtered matrices.
}
\description{
Generic to access eQTLObject filter data
}
\examples{
  data(testEQTL)
  get_filter_data(testEQTL)

}
