% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model_info,eQTLObject-method}
\alias{get_model_info,eQTLObject-method}
\title{Method to access eQTLObject used model information}
\usage{
\S4method{get_model_info}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
used model information of eQTLObject.
}
\description{
Method to access eQTLObject used model information
}
