% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model_info}
\alias{get_model_info}
\title{Generic to access eQTLObject used model information}
\usage{
get_model_info(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
used model information of eQTLObject.
}
\description{
Generic to access eQTLObject used model information
}
\examples{
  data(testEQTL)
  get_model_info(testEQTL)

}
