% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_raw_data,eQTLObject-method}
\alias{get_raw_data,eQTLObject-method}
\title{Method to access eQTLObject raw data}
\usage{
\S4method{get_raw_data}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
raw data matrix.
}
\description{
Method to access eQTLObject raw data
}
