% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Generic to access eQTLObject raw data}
\usage{
get_raw_data(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
raw data matrix.
}
\description{
Generic to access eQTLObject raw data
}
\examples{
  data(testEQTL)
  get_raw_data(testEQTL)

}
