% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_result_info,eQTLObject-method}
\alias{get_result_info,eQTLObject-method}
\title{Method to access the result of identifying eQTLs from scRNA-seq data}
\usage{
\S4method{get_result_info}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
A dataframe.
}
\description{
Method to access the result of identifying eQTLs from scRNA-seq data
}
