% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_normalize-normalizeGene.R
\name{limma_normalize}
\alias{limma_normalize}
\title{Normalize the gene expression matrix with limma}
\usage{
limma_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{A numeric matrix of raw gene expression counts,
with genes as rows and cells as columns.}
}
\value{
A normalized gene expression matrix after applying limma
normalization.
}
\description{
\code{limma_normalize()} normalizes an expression matrix using the quantile
normalization method provided by the limma package.
}
\examples{
data(GeneData)
limma_normalize(GeneData)
}
