% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_biclassify_info,eQTLObject-method}
\alias{load_biclassify_info,eQTLObject-method}
\title{Method to access eQTLObject biclassify information}
\usage{
\S4method{load_biclassify_info}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
biclassify information of eQTLObject.
}
\description{
Method to access eQTLObject biclassify information
}
