% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_biclassify_info}
\alias{load_biclassify_info}
\title{Generic to access eQTLObject biclassify information}
\usage{
load_biclassify_info(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
biclassify information of eQTLObject.
}
\description{
Generic to access eQTLObject biclassify information
}
\examples{
  data(testEQTL)
  load_biclassify_info(testEQTL)

}
