% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_group_info,eQTLObject-method}
\alias{load_group_info,eQTLObject-method}
\title{Method to access eQTLObject cell grouping information}
\usage{
\S4method{load_group_info}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
A dataframe.
}
\description{
Method to access eQTLObject cell grouping information
}
