% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_group_info}
\alias{load_group_info}
\title{Generic to access eQTLObject cell grouping information}
\usage{
load_group_info(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
A dataframe.
}
\description{
Generic to access eQTLObject cell grouping information
}
\examples{
  data(testEQTL)
  load_group_info(testEQTL)

}
