% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_species_info,eQTLObject-method}
\alias{load_species_info,eQTLObject-method}
\title{Method to access eQTLObject species information}
\usage{
\S4method{load_species_info}{eQTLObject}(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
species information of eQTLObject.
}
\description{
Method to access eQTLObject species information
}
