% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_theme_opts.R
\name{plots_theme_opts}
\alias{plots_theme_opts}
\title{Customized ggplot2 Theme for Plots}
\usage{
plots_theme_opts()
}
\value{
A \code{ggplot2} theme object for styling plots.
}
\description{
Customized ggplot2 Theme for Plots
}
\examples{
library(ggplot2)
data <- data.frame(
  x = c("A", "B", "C", "D", "E"),
  y = c(10, 20, 30, 40, 50))
ggplot(data, aes(x, y)) +
  geom_point() +
  plots_theme_opts()
}
