% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQTLtools-package.R
\docType{package}
\name{scQTLtools-package}
\alias{scQTLtools}
\alias{scQTLtools-package}
\title{scQTLtools: An R package for single-cell eQTL analysis and visualization}
\description{
This package specializes in analyzing and visualizing eQTL at the single-cell level. It can read gene expression matrices or Seurat data, or SingleCellExperiment object along with genotype data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XFWuCN/scQTLtools}
  \item Report bugs at \url{https://github.com/XFWuCN/scQTLtools/issues}
}

}
\author{
\strong{Maintainer}: Xiaofeng Wu \email{1427972815@qq.com} (\href{https://orcid.org/0009-0003-6254-5575}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Xin Huang \email{1097567240@qq.com} (\href{https://orcid.org/0009-0005-2755-0357}{ORCID}) [copyright holder]
  \item Siwen Xu \email{siwxu@gdpu.edu.cn} (\href{https://orcid.org/0000-0001-7936-0639}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Jingtong Kang \email{1203178107@qq.com} (\href{https://orcid.org/0009-0008-8343-3456}{ORCID}) [compiler]
}

}
\keyword{internal}
