% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_filter_data}
\alias{set_filter_data}
\title{Generic to set eQTLObject filter data}
\usage{
set_filter_data(x, value, name)
}
\arguments{
\item{x}{A eQTLObject object.}

\item{value}{The filtered data.}

\item{name}{The matrix named 'name' is stored under the 'filterData' slot as
an element within its list.}
}
\value{
eQTLObject.
}
\description{
Generic to set eQTLObject filter data
}
\examples{
  data(testEQTL)
  data123 <- matrix(0, nrow = 3, ncol = 3)
  set_filter_data(testEQTL, data123, "expMat")

}
