% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_model_info,eQTLObject-method}
\alias{set_model_info,eQTLObject-method}
\title{Method to set eQTLObject used model information}
\usage{
\S4method{set_model_info}{eQTLObject}(x, value)
}
\arguments{
\item{x}{A eQTLObject object.}

\item{value}{The used model information to set to eQTLObject.}
}
\value{
eQTLObject.
}
\description{
Method to set eQTLObject used model information
}
\examples{
  data(testEQTL)
  useModel <- "zinb"
  set_model_info(testEQTL, useModel)

}
