% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_model_info}
\alias{set_model_info}
\title{Generic to set eQTLObject used model information}
\usage{
set_model_info(x, value)
}
\arguments{
\item{x}{A eQTLObject object.}

\item{value}{The used model information to set to eQTLObject.}
}
\value{
eQTLObject.
}
\description{
Generic to set eQTLObject used model information
}
\examples{
  data(testEQTL)
  useModel <- "zinb"
  set_model_info(testEQTL, useModel)

}
