% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_raw_data,eQTLObject-method}
\alias{set_raw_data,eQTLObject-method}
\title{Method to set eQTLObject raw data}
\usage{
\S4method{set_raw_data}{eQTLObject}(x, value, name)
}
\arguments{
\item{x}{A eQTLObject object.}

\item{value}{The raw data.}

\item{name}{The matrix named 'name' is stored under the 'rawData' slot as an
element within its list.}
}
\value{
eQTLObject.
}
\description{
Method to set eQTLObject raw data
}
\examples{
  data(testEQTL)
  data123 <- matrix(0, nrow = 3, ncol = 3)
  set_raw_data(testEQTL, data123, "rawExpMat")

}
