% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_result_info,eQTLObject-method}
\alias{set_result_info,eQTLObject-method}
\title{Method to set the result of identifying eQTLs from scRNA-seq data}
\usage{
\S4method{set_result_info}{eQTLObject}(x, value)
}
\arguments{
\item{x}{A eQTLObject object.}

\item{value}{A dataframe, each row describes eQTL discovering result of a
SNP-Gene pair.}
}
\value{
eQTLObject.
}
\description{
Method to set the result of identifying eQTLs from scRNA-seq data
}
\examples{
  data(testEQTL)
  result <- matrix(0, nrow = 3, ncol = 3)
  set_result_info(testEQTL, result)

}
