% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM_normalize-normalizeGene.R
\name{TPM_normalize}
\alias{TPM_normalize}
\title{Normalize the gene expression matrix with TPM}
\usage{
TPM_normalize(expressionMatrix)
}
\arguments{
\item{expressionMatrix}{A numeric matrix of raw gene expression counts,
with genes as rows and cells as columns.}
}
\value{
A normalized gene expression matrix after applying TPM
normalization.
}
\description{
\code{TPM_normalize()} scales an expression matrix using Transcripts Per
Million (TPM) normalization, applying logarithm and scaling operations to
adjust data based on library size.
}
\examples{
data(GeneData)
TPM_normalize(GeneData)
}
