% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_by_id-utils.R
\name{get_cell_groups}
\alias{get_cell_groups}
\title{Retrieve Cells by SNP Value}
\usage{
get_cell_groups(snpMatrix, SNPid, biClassify)
}
\arguments{
\item{snpMatrix}{A matrix containing SNP data where rows represent SNPs
and columns represent cells.}

\item{SNPid}{A character string or numeric index representing the specific
SNP of interest in the SNP matrix.}

\item{biClassify}{The user chooses whether to convert the counting method of
the snpMatrix to 0/1/2, TRUE indicates conversion,
and FALSE indicates no conversion, default is no conversion.}
}
\value{
A list of cell names (column names of the SNP matrix)
        that correspond to the specified genotype value for the given SNP.
}
\description{
This function extracts the names of cells from a SNP matrix that correspond
to a specified value for a given SNP.
}
\examples{
data(testSNP)
biClassify <- FALSE
get_cell_groups(testSNP, "1:632445", biClassify)
}
