% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model_info}
\alias{get_model_info}
\alias{get_model_info,eQTLObject-method}
\title{Access model specification from an eQTLObject}
\usage{
get_model_info(x)

\S4method{get_model_info}{eQTLObject}(x)
}
\arguments{
\item{x}{An \code{eQTLObject}.}
}
\value{
A character string indicating the model.

used model information of eQTLObject.
}
\description{
Access model specification from an eQTLObject

Method to access eQTLObject used model information.
}
\examples{
  data(testEQTL)
  get_model_info(testEQTL)

}
