% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_result_info}
\alias{get_result_info}
\alias{get_result_info,eQTLObject-method}
\title{Access eQTLs results from an eQTLObject.}
\usage{
get_result_info(x)

\S4method{get_result_info}{eQTLObject}(x)
}
\arguments{
\item{x}{An \code{eQTLObject}.}
}
\value{
A data frame where each row corresponds to an identified gene-SNP
pair.

A data frame with eQTL results.
}
\description{
Access eQTLs results from an eQTLObject.

Method to access the result of identifying eQTLs.
}
\examples{
  data(testEQTL)
  get_result_info(testEQTL)

}
