% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_group_info}
\alias{load_group_info}
\alias{load_group_info,eQTLObject-method}
\title{Access cell grouping information from an eQTLObject}
\usage{
load_group_info(x)

\S4method{load_group_info}{eQTLObject}(x)
}
\arguments{
\item{x}{An \code{eQTLObject}.}
}
\value{
A data frame with grouping information.

A data frame with cell group assignments.
}
\description{
Access cell grouping information from an eQTLObject

Method to access eQTLObject cell grouping information.
}
\examples{
  data(testEQTL)
  load_group_info(testEQTL)

}
