% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_species_info}
\alias{load_species_info}
\title{Generic to access eQTLObject species information}
\usage{
load_species_info(x)
}
\arguments{
\item{x}{A eQTLObject object.}
}
\value{
species information of eQTLObject.
}
\description{
Generic to access eQTLObject species information
}
\examples{
  data(testEQTL)
  load_species_info(testEQTL)

}
