% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_theme_opts.R
\name{plots_theme_opts}
\alias{plots_theme_opts}
\title{Theme options for customized plots}
\usage{
plots_theme_opts()
}
\value{
A ggplot2 theme object with customized settings.
}
\description{
Theme options for customized plots
}
\examples{
library(ggplot2)
data <- data.frame(
  x = c("A", "B", "C", "D", "E"),
  y = c(10, 20, 30, 40, 50))
ggplot(data, aes(x, y)) +
  geom_point() +
  plots_theme_opts()
}
